bump_version.py: bump version in conda env export
At first, I added pattern matches for each of the occurences in the
Dockerfile, however this looked kinda ugly because there were multiple
'Updating version in Dockerfile ..' lines printed.

So I refactored the code a little to make sure there is only one logging.info line printed per file that was changed.
While I was at it, I also added a bit of color to the output.