test: Ignore pipeline name line
When the template is generated the description and name take up more
than 120 characters, so then prettier re-formats the email_template and
breaks it up into 3 separate lines because it's longer the 120.

We can't just break it up into 3 lines though because if there's a short
name and description it'll reformat it into one line.

Co-authored-by: Rike <friederike.hanssen@qbic.uni-tuebingen.de>