Make cli-provided module/subworkflow names case insensitive
Uses a click callback function to normalize the name of a module/subworkflow
that is provided by the user on the CLI, so that names written in uppercase
still resolve to the expected lowercase name.

In case no argument is provided, the callback function simply returns None,
and thus the current behaviour of a user being asked to select a component
via an interactive prompt is maintained.

The .casefold() method is used in favour of .lower() for slightly
more robust case normalization.
See: https://stackoverflow.com/questions/45745661/lower-vs-casefold-in-string-matching-and-converting-to-lowercase

The callback is applied to all occurrences of "tool/module/subworkflow" arguments,
except those in the the "create" functions, because these already contain an internal
check that prompts to user to adhere to the lowercase naming convention.

Updates to CHANGELOG.md.

Co-authored-by: Adam Talbot <adam.talbot@seqera.io>
Co-authored-by: Phil Ewels <phil.ewels@seqera.io>
Co-authored-by: Arthur Gymer <24782660+awgymer@users.noreply.github.com>