Add linkify-py-it as explicit dependency
The bioconda version of nf-core tools produced errors
when running commands that launched the trogon TUI.
The reason was the missing package linkify,
which gets pulled in as a dependency of textual in the PyPi
build, but not in the conda-forge recipe.
See https://github.com/Textualize/textual/blob/22770300252deb28d266fe4ed4766d6e2a2f5dd2/pyproject.toml#L44,
https://github.com/conda-forge/textual-feedstock/blob/main/recipe/meta.yaml
and nf-core/tools#3257.