Refactor generate_downstream_samplesheets subworkflow
This commit refactors the generate_downstream_samplesheets subworkflow in the pipeline-template. It updates the workflow to a specific pipeline and adds support for more pipelines in the future. The changes include modifying the input parameters and updating the channelToSamplesheet function. This refactoring improves the flexibility and extensibility of the subworkflow.