Remove toList() channel operation from inside onComplete block
This PR resolves an important bug in the nf-core template, whereby all workflows will hang if the --email parameter is supplied.

The onComplete block will hang if there are any (queue) channel operations inside the block. All values in the onComplete block must be resolved to single values or value channels _before_ the onComplete block starts.

The async channels are not available inside onComplete, so calling the toList() operation will hang forever as the async queue channel will never be completed.