fix(pipeline-sync): Removes requirement to have authors defined
If authors were missing the pipelines sync failed with a meaningless error. This wastes developer time while users have to fight configuration. Since authors are easy to maintain manually it seems silly to break the pipeline sync just for on field, so this PR makes it optional, it is encouraged by the .nf-core.yml and existing template.