exit code for unrecoverable snapshots errors
When a pipeline process is in the process of being snapshotted via Fusion Snapshots
and encounters an unrecoverable error the Fusion Snapshots runtime emits
a specific exit code (175) to signal that to Nextflow.

This PR makes sure that users creating pipelines out of the template
are including the exit code in their retry strategy.