Fix pipeline_if_empty_null long runtime
This borrows the logic from pipeline_todos for ignoring all paths in .gitignore
when walking the root dir, which I believe fixes #3605.  It also changes to
calling `findall` directly on the precompiled Pattern object rather than
calling `re.findall(pattern, ...)` as the latter incurs a non-trivial overhead
to check the internal pattern cache.