Improve file ignores in workflow file enumeration
utils.get_wf_files has logic for parsing .gitignore files and using the
patterns there as an ignore list, but it doesn't implement the full semantics
of gitignore patterns. This change adds support for trailing slash which should
ignore an entire directory, which is important as such patterns are included in
the pipeline TEMPLATE .gitignore.

Also, relative path handling is implemented in the ignore pattern checking.
This means ignore patterns still work correctly when the -d option is given to
lint a directory other than the current one.

Finally, .git/* is added as an implicit ignore pattern that is always in
effect.